feat(discord): organise and clean up todos and comments
