+++ integrations/discord/src/index.tsx
@@ -9,6 +9,7 @@ import {
     EventCallback,
 } from '@gitbook/runtime';
 
+import { sendDiscordMessage } from './discord';
 import { DiscordRuntimeContext, OAuthResponseWebhook } from './types';
 
 /*
@@ -18,40 +19,7 @@ const handleSpaceContentUpdated: EventCallback<
     'space_content_updated',
     DiscordRuntimeContext
 > = async (event, context) => {
-    const { environment, api } = context;
-    const { data: space } = await api.spaces.getSpaceById(event.spaceId);
-
-    const accessToken = environment.installation?.configuration.oauth_credentials?.access_token;
-    const botToken = environment.secrets.BOT_TOKEN;
-
-    if (!accessToken) {
-        throw new Error('No authentication token provided');
-    }
-
-    const channelId =
-        environment.installation?.configuration.oauth_credentials?.webhook?.channel_id;
-
-    const embedObject = {
-        title: 'View changes here:',
-        description: `${space.urls.app}`,
-        thumbnail: {
-            url: 'https://avatars.githubusercontent.com/u/7111340?s=280&v=4',
-        },
-    };
-
-    const url = `https://discord.com/api/v10/channels/${channelId}/messages`;
-    const headers = {
-        'Content-Type': 'application/json',
-        Authorization: `Bot ${botToken}`,
-    };
-    const body = JSON.stringify({
-        content: `Changes have been published in ${space.title}.`,
-        embeds: [embedObject],
-    });
-
-    await fetch(url, { method: 'POST', headers, body }).catch((err) => {
-        throw new Error(`Error fetching content from ${url}. ${err}`);
-    });
+    await sendDiscordMessage(event, context);
 };
 
 const handleFetchEvent: FetchEventCallback<DiscordRuntimeContext> = async (request, context) => {
@@ -111,7 +79,6 @@ const extractCredentials = async (
 export default createIntegration<DiscordRuntimeContext>({
     fetch: handleFetchEvent,
     events: {
-        // TODO: Update this event to the correct one for when a change is merged - I only used this one as it is quicker to test the changes
         space_content_updated: handleSpaceContentUpdated,
     },
 });
